+++ src/Controller/UserController.php
@@ -2,9 +2,18 @@
 
 namespace App\Controller;
 
+use App\Entity\Tricks;
+use App\Entity\User;
+
+use App\Form\TricksType;
+
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Doctrine\Persistence\ManagerRegistry;
+use Symfony\Component\HttpFoundation\File\Exception\FileException;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\String\Slugger\SluggerInterface;
 
 /**
  * UserController
@@ -16,6 +25,63 @@ class UserController extends AbstractController
     {
         return $this->render('user/index.html.twig', [
             'controller_name' => 'UserController',
+        ]);
+    }
+
+    /**
+     * function addTricks
+     * 
+     * @return void
+     */
+    #[Route('/tricks/add', name: 'app_user_tricks_add')]
+    public function addTricks(Request $request, ManagerRegistry $doctrine, SluggerInterface $slugger): Response
+    {
+        $tricks =  new Tricks();
+        $formAddTrick = $this->createForm(TricksType::class, $tricks);
+        $formAddTrick->handleRequest($request);
+
+        if ($formAddTrick->isSubmitted() && $formAddTrick->isValid()) {
+            $pictureFile =  $formAddTrick->get('picture')->getData();
+
+            if ($pictureFile) {
+                $originalFilename = $pictureFile;
+                $safeFilename = $slugger->slug($originalFilename);  // not used
+                $newFilename = md5(uniqid()) . '.' . $originalFilename->guessExtension();
+                $tricks->setPicture($newFilename);
+                try {
+                    $pictureFile->move(
+                        $this->getParameter('pictues_directory'),
+                        $newFilename
+                    );
+                } catch (FileException $e) {
+                    // ... handle exception if something happens during file upload
+                    //soit redirigé sur la page du formulaire en cas d'erreur, en précisant le(s) type(s) d'erreurs ;
+
+                    $message = $this->addFlash('error', 'error type:' . $e); // or in twig
+
+                    return $this->render('tricks/add.html.twig', [
+                        'formAddTrick' =>  $formAddTrick->createView(),
+                        'message' => $message
+                    ]);
+                }
+                $tricks->setPicture($newFilename);
+            }
+
+            //    $txt = $formAddTrick->get('content')->getData();  to see later     
+
+            $tricks->setUser($this->getUser());
+
+            $entityManager = $doctrine->getManager();
+            $entityManager->persist($tricks);
+            $entityManager->flush();
+
+            $this->addFlash('success', 'Your trick have been added.');
+
+            return $this->redirectToRoute('app_home');
+        } //end form
+
+        return $this->render('tricks/add.html.twig', [
+            'formAddTrick' =>  $formAddTrick->createView(),
         ]);
     }
 }
